ARM: NUMA: DT: Do not expose numa info to DOM0
Delete numa-node-id and distance map from DOM0 DT
so that NUMA information is not exposed to DOM0.
This helps particularly to boot Node 1 devices
as if booting on Node0.

However this approach has limitation where memory allocation
for the devices should be local.

Also, do not expose numa distance node to DOM0.

Signed-off-by: Vijaya Kumar <Vijaya.Kumar@cavium.com>